refactor(crud): refactor delete request
