+++ jina/types/request/control.py
@@ -1,10 +1,7 @@
 from . import Request
 
-from ...proto import jina_pb2
+from .mixin import CommandMixin
 
 
-class ControlRequest(Request):
-    @property
-    def command(self) -> str:
-        self.is_used = True
-        return jina_pb2.RequestProto.ControlRequestProto.Command.Name(self.as_pb_object.control.command)
+class ControlRequest(Request, CommandMixin):
+    pass
