+++ jina/types/request/mixin.py
@@ -1,4 +1,5 @@
 from ..sets import DocumentSet
+from ...proto import jina_pb2
 
 
 class DocsPropertyMixin:
@@ -13,3 +14,16 @@ class GroundtruthPropertyMixin:
     def groundtruths(self) -> 'DocumentSet':
         self.is_used = True
         return DocumentSet(self.body.groundtruths)
+
+
+class IdsMixin:
+    @property
+    def ids(self):
+        return self.body.ids
+
+
+class CommandMixin:
+    @property
+    def command(self) -> str:
+        self.is_used = True
+        return jina_pb2.RequestProto.ControlRequestProto.Command.Name(self.as_pb_object.control.command)
